Test: RED ->test infrastructure 1st ver finished

+++ test/lcd_hd44780/mock_LCD_IO_driver.h
@@ -2,7 +2,7 @@
  * @Author: lukasz.niewelt 
  * @Date: 2023-12-06 22:11:42 
  * @Last Modified by: lukasz.niewelt
- * @Last Modified time: 2023-12-06 23:58:24
+ * @Last Modified time: 2023-12-07 00:13:30
  */
 #include <stdint.h>
 #define BUF_SIZE 1000
@@ -21,7 +21,7 @@ extern uint8_t mock_LCD_SIG_PORT;
 extern uint8_t mock_LCD_DATA_PORT_DIRECTION;
 extern uint8_t mock_LCD_SIG_PORT_DIRECTION;
 
-extern uint32_t mock_LCD_Port_delay_dump_data[BUF_SIZE][LOG_DATA_AMOUNT];
+extern uint16_t mock_LCD_Port_delay_dump_data[BUF_SIZE][LOG_DATA_AMOUNT];
 
 uint8_t mock_get_lcd_init_state(void);
